Start splitting the linter into smaller files
